[tvla] Add bit-based histogram function
Move histogram functions into a separate file
Add compute_histograms_aes_bit
