+++ analysis/tvla.py
@@ -23,6 +23,8 @@ sys.path.append(ABS_PATH + '/..')
 import util.plot as plot  # noqa : E402
 from capture.project_library.project import ProjectConfig  # noqa : E402
 from capture.project_library.project import SCAProject  # noqa : E402
+from util.histograms import compute_histograms_aes_byte  # noqa : E402
+from util.histograms import compute_histograms_general  # noqa : E402
 from util.leakage_models import compute_leakage_aes_byte  # noqa : E402
 from util.leakage_models import compute_leakage_general  # noqa : E402
 from util.leakage_models import find_fixed_entry  # noqa : E402
@@ -112,55 +114,6 @@ def compute_statistics(num_orders, rnd_list, byte_list, histograms, x_axis):
             ttest_trace[:, i_rnd, i_byte, :] = tmp
 
     return ttest_trace
-
-
-def compute_histograms_general(trace_resolution, traces, leakage):
-    """ Building histograms for general fixed-vs-random TVLA.
-
-    For each time sample we make two histograms, one for the fixed and one for the random group.
-    Whether a trace belongs to the fixed or random group is indicated in the leakage input
-    variable. The value stored in histograms[v][w][x][y][z] shows how many traces have value z at
-    time y, given that trace is in the fixed (x = 1) or random (x = 0) group. The v and w indices
-    are not used but we keep them for code compatiblitly with non-general AES TVLA.
-    """
-    num_leakages = 2
-    num_rnds = 1
-    num_bytes = 1
-    num_samples = traces.shape[1]
-    histograms = np.zeros((num_rnds, num_bytes, num_leakages, num_samples, trace_resolution),
-                          dtype=np.uint32)
-
-    for i_sample in range(num_samples):
-        histograms[0, 0, :, i_sample, :] = np.histogram2d(
-            leakage, traces[:, i_sample],
-            bins=[range(num_leakages + 1), range(trace_resolution + 1)])[0]
-
-    return histograms
-
-
-def compute_histograms_aes(trace_resolution, rnd_list, byte_list, traces, leakage):
-    """ Building histograms for AES.
-
-    For each time sample we make two histograms, one for Hamming weight of the sensitive variable
-    = 0 (fixed set) and one for Hamming weight > 0 (random set). The value stored in
-    histograms[v][w][x][y][z] shows how many traces have value z at time y, given that
-    HW(state byte w in AES round v) = 0 (fixed set, x = 0) or > 0 (random set, x = 1).
-    """
-    num_leakages = 2
-    num_rnds = len(rnd_list)
-    num_bytes = len(byte_list)
-    num_samples = traces.shape[1]
-    histograms = np.zeros((num_rnds, num_bytes, num_leakages, num_samples, trace_resolution),
-                          dtype=np.uint32)
-
-    for i_rnd in range(num_rnds):
-        for i_byte in range(num_bytes):
-            for i_sample in range(num_samples):
-                histograms[i_rnd, i_byte, :, i_sample, :] = np.histogram2d(
-                    leakage[rnd_list[i_rnd], byte_list[i_byte], :], traces[:, i_sample],
-                    bins=[np.append(range(num_leakages), 9), range(trace_resolution + 1)])[0]
-
-    return histograms
 
 
 def tvla_plotting_fnc(axs, num_orders, i_rnd, i_byte, ttest_trace,
@@ -652,8 +605,8 @@ def run_tvla(ctx: typer.Context):
                 # > 0 (random set, x = 1).
                 # The computation is parallelized over the samples.
                 histograms = Parallel(n_jobs=num_jobs)(
-                    delayed(compute_histograms_aes)(trace_resolution, rnd_list, byte_list,
-                                                    traces[:, i:i + sample_step_hist], leakage)
+                    delayed(compute_histograms_aes_byte)(trace_resolution, rnd_list, byte_list,
+                                                         traces[:, i:i + sample_step_hist], leakage)
                     for i in range(0, num_samples, sample_step_hist))
                 histograms = np.concatenate((histograms[:]), axis=3)
             else:
